pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "k8s-ai-support"
version = "1.0.0"
description = "AI-powered Kubernetes troubleshooting agent with CLI and MCP server"
authors = ["k8s-ai-support <dev@k8s-ai-support.io>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/your-org/k8s-ai-support"
repository = "https://github.com/your-org/k8s-ai-support"
keywords = ["kubernetes", "k8s", "troubleshooting", "ai", "llm", "mcp", "devops", "sre"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "src"}]

[tool.poetry.scripts]
k8s-ai-support = "src.cli.main:main"
kas = "src.cli.main:main"

[project.scripts]
k8s-ai-support = "src.cli.main:main"
kas = "src.cli.main:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

# Core LangChain / LangGraph
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langgraph = "^0.2.0"

# LLM Provider SDKs (install the ones you need)
langchain-openai = {version = "^0.2.0", optional = true}
langchain-google-genai = {version = "^2.0.0", optional = true}
langchain-anthropic = {version = "^0.3.0", optional = true}
langchain-community = {version = "^0.3.0", optional = true}  # For Ollama

# RAG / Embeddings
sentence-transformers = {version = "^3.0.0", optional = true}  # Local embeddings (no API key)
chromadb = {version = "^0.5.0", optional = true}               # Vector store

# Token counting
tiktoken = {version = "^0.8.0", optional = true}

# CLI
typer = {version = "^0.12.0", extras = ["all"]}
rich = "^13.7.0"

# Settings / validation
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"

# Async
anyio = "^4.4.0"
nest-asyncio = "^1.6.0"

# Logging
structlog = "^24.4.0"

# HTTP (for Ollama and LLM providers)
httpx = "^0.27.0"

# YAML output support
pyyaml = "^6.0.2"

[tool.poetry.extras]
openai = ["langchain-openai", "tiktoken"]
gemini = ["langchain-google-genai"]
claude = ["langchain-anthropic"]
ollama = ["langchain-community"]
rag = ["sentence-transformers", "chromadb"]
all = [
    "langchain-openai",
    "langchain-google-genai",
    "langchain-anthropic",
    "langchain-community",
    "sentence-transformers",
    "chromadb",
    "tiktoken",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
mypy = "^1.11.0"
ruff = "^0.6.0"
black = "^24.8.0"
pre-commit = "^3.8.0"
ipython = "^8.26.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=70",
    "-v",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/conftest.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in default args
    "B024",   # abstract base class with no abstract methods
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
